[torchgen][RFC] Refactor types and generate Executorch code
In `types_base`:
* Created base class `CType`. `BaseCType` and `ConstRefCType` etc are inheriting `CType`.
* Only keep abstract type model definitions, such as `BaseCppType`.

In `types`:
* Define `BaseCppType` with `at` and `c10` namespaces.
* All the signatures using these types.

Created a module for Executorch: `torchgen.executorch`.

In `torchgen.executorch.api.types`:
* Define `BaseCppType` with `torch::executor` namespace.
* All the signatures using these types. (`CppSignature` and `NativeSignature`)

In `torchgen.executorch.api.cpp`:
* Help generate `NamedCType` for `CppSignature` arguments.

 In `torchgen.executorch.api.native`:
* Help generate `NamedCType` for `NativeSignature` arguments.

Differential Revision: [D41455634](https://our.internmc.facebook.com/intern/diff/D41455634/)

[ghstack-poisoned]
