+++ .flake8
@@ -12,7 +12,7 @@ ignore =
     B007,B008,
     # these ignores are from flake8-comprehensions; please fix!
     C400,C401,C402,C403,C404,C405,C407,C411,C413,C414,C415
-per-file-ignores = __init__.py: F401 torch/utils/cpp_extension.py: B950
+per-file-ignores = __init__.py: F401 torch/utils/cpp_extension.py: B950 torchgen/api/types/__init__.py: F401,F403
 optional-ascii-coding = True
 exclude =
     ./.git,
